Organize code/callbacks
Separate callbacks in a different file for readability and better organization. Also creates a wrapper.py file that calls the TLS handshake parsing (wrapper not implemented yet).
